renamed wildfire dashboard map & chart components
