+++ src/routes/tools/wildfire/_WildfireChart.svelte
